:recycle: MyVideo와 OtherVideo 별도의 컴포넌트로 분리
